Fixes and tests
